Delete, and view employee details 🚶🏻‍♂️

+++ src/main/java/com/dfms/dairy_farm_management_system/controllers/EmployeesController.java
@@ -1,18 +1,22 @@
 package com.dfms.dairy_farm_management_system.controllers;
 
-import com.dfms.dairy_farm_management_system.connection.DBConfig;
+import com.dfms.dairy_farm_management_system.Main;
+import com.dfms.dairy_farm_management_system.controllers.pop_ups_controllers.EmployeeDetailsController;
 import com.dfms.dairy_farm_management_system.models.Employee;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.geometry.Insets;
+import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.HBox;
+import javafx.stage.Stage;
 import javafx.util.Callback;
 
 import java.io.IOException;
@@ -24,12 +28,13 @@ import java.sql.Statement;
 import java.util.Optional;
 import java.util.ResourceBundle;
 
-import static com.dfms.dairy_farm_management_system.helpers.Helper.displayAlert;
-import static com.dfms.dairy_farm_management_system.helpers.Helper.openNewWindow;
+import static com.dfms.dairy_farm_management_system.connection.DBConfig.getConnection;
+import static com.dfms.dairy_farm_management_system.helpers.Helper.*;
 
 public class EmployeesController implements Initializable {
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
+        ObservableList<String> list = FXCollections.observableArrayList("PDF", "Excel");
         export_cb.setItems(list);
         displayEmployees();
         liveSearch(search_employee_input, employees_table);
@@ -37,7 +42,8 @@ public class EmployeesController implements Initializable {
 
     private Statement st;
     private PreparedStatement pst;
-    private Connection con = DBConfig.getConnection();
+    private Connection con = getConnection();
+
     @FXML
     private TableView<Employee> employees_table;
     @FXML
@@ -68,7 +74,6 @@ public class EmployeesController implements Initializable {
 
     @FXML
     private Button openAddNewEmployeeBtn;
-    ObservableList<String> list = FXCollections.observableArrayList("PDF", "Excel");
 
     //get all the employees
     public ObservableList<Employee> getEmployees() {
@@ -162,18 +167,46 @@ public class EmployeesController implements Initializable {
 
                         setGraphic(managebtn);
                         setText(null);
+
+                        //action for action buttons
                         delete_btn.setOnMouseClicked((MouseEvent event) -> {
-                            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-                            alert.setTitle("Delete Employee");
-                            alert.setHeaderText(null);
-                            alert.setContentText("Are you sure you want to delete this employee?");
-                            Optional<ButtonType> action = alert.showAndWait();
-                            if (action.get() == ButtonType.OK) {
-                                //get selected item of clicked button
-                                Employee employee = getTableView().getItems().get(getIndex());
-                                deleteEmployee(employee.getId());
-                                displayEmployees();
+                            Employee employee = employees_table.getSelectionModel().getSelectedItem();
+                            if (employee != null) {
+                                Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                                alert.setTitle("Confirmation");
+                                alert.setHeaderText("Are you sure you want to delete this employee?");
+                                alert.setContentText("Click ok to confirm");
+                                Optional<ButtonType> action = alert.showAndWait();
+                                if (action.get() == ButtonType.OK) {
+                                    deleteEmployee(employee.getId());
+                                    displayEmployees();
+                                }
+                            } else {
+                                displayAlert("Error", "Please select an employee to delete", Alert.AlertType.ERROR);
                             }
+                        });
+
+                        view_details_btn.setOnMouseClicked((MouseEvent event) -> {
+                            int id = employees_table.getSelectionModel().getSelectedItem().getId();
+                            String url = "popups/employee_details.fxml";
+                            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/com/dfms/dairy_farm_management_system/popups/employee_details.fxml"));
+                            Scene scene = null;
+                            try {
+                                scene = new Scene(fxmlLoader.load());
+                                EmployeeDetailsController controller = fxmlLoader.getController();
+                                System.out.println("id: " + id);
+                                controller.setEmployeeId(id);
+                            } catch (IOException e) {
+                                displayAlert("Error", e.getMessage(), Alert.AlertType.ERROR);
+                                e.printStackTrace();
+                            }
+                            Stage stage = new Stage();
+                            stage.getIcons().add(new Image("file:src/main/resources/images/logo.png"));
+                            stage.setTitle("Employee Details");
+                            stage.setResizable(false);
+                            stage.setScene(scene);
+                            centerScreen(stage);
+                            stage.show();
                         });
                     }
                 }
@@ -227,4 +260,41 @@ public class EmployeesController implements Initializable {
     void searchEmployee(MouseEvent event) {
         liveSearch(this.search_employee_input, employees_table);
     }
+
+    public void displayEmployeeConsole(Employee employee) {
+        System.out.println("Employee ID: " + employee.getId());
+        System.out.println("Employee First Name: " + employee.getFirstName());
+        System.out.println("Employee Last Name: " + employee.getLastName());
+        System.out.println("Employee Email: " + employee.getEmail());
+        System.out.println("Employee Phone: " + employee.getPhone());
+        System.out.println("Employee Address: " + employee.getAdress());
+        System.out.println("Employee Cin: " + employee.getCin());
+        System.out.println("Employee gender: " + employee.getGender());
+        System.out.println("Employee Recrutement Date: " + employee.getRecruitmentDate());
+    }
+
+//    public Employee getEmployee(int id) {
+//        Employee employee = new Employee();
+//        String query = "SELECT * FROM employee WHERE id = " + id;
+//        con = getConnection();
+//        try {
+//            st = con.createStatement();
+//            ResultSet rs = st.executeQuery(query);
+//            while (rs.next()) {
+//                employee.setId(rs.getInt("id"));
+//                employee.setFirstName(rs.getString("first_name"));
+//                employee.setLastName(rs.getString("last_name"));
+//                employee.setEmail(rs.getString("email"));
+//                employee.setPhone(rs.getString("phone"));
+//                employee.setAdress(rs.getString("address"));
+//                employee.setCin(rs.getString("cin"));
+//                employee.setGender(rs.getString("gender"));
+//                employee.setRecruitmentDate(rs.getDate("recruitment_date"));
+//                employee.setSalary(rs.getFloat("salary"));
+//            }
+//        } catch (Exception e) {
+//            displayAlert("Error", e.getMessage(), Alert.AlertType.ERROR);
+//        }
+//        return employee;
+//    }
 }
